+++ src/gui/widgets/dialogs/import_export_dialog.hpp
@@ -39,6 +39,7 @@ public:
         dialog.setWindowTitle(QObject::tr("Save file"));
         dialog.setAcceptMode(QFileDialog::AcceptSave);
         dialog.setFileMode(QFileDialog::AnyFile);
+        dialog.setOption(QFileDialog::DontUseNativeDialog, !app::settings::get<bool>("open_save", "native_dialog"));
         setup_file_dialog(dialog, io::IoRegistry::instance().exporters(), io_options_.format, false);
         while ( true )
         {
@@ -81,6 +82,7 @@ public:
         dialog.setWindowTitle(QObject::tr("Open file"));
         dialog.setAcceptMode(QFileDialog::AcceptOpen);
         dialog.setFileMode(QFileDialog::ExistingFile);
+        dialog.setOption(QFileDialog::DontUseNativeDialog, !app::settings::get<bool>("open_save", "native_dialog"));
         setup_file_dialog(dialog, io::IoRegistry::instance().importers(), io_options_.format, true);
         if ( show_file_dialog(dialog, io::IoRegistry::instance().importers(), io::ImportExport::Import) )
             return options_dialog(io_options_.format->open_settings());
@@ -111,6 +113,8 @@ private:
         dialog.setOption(QFileDialog::DontUseNativeDialog, false);
         dialog.setFixedSize(dialog.parentWidget()->size());
         dialog.setWindowState(dialog.windowState() | Qt::WindowFullScreen);
+#else
+        dialog.setOption(QFileDialog::DontUseNativeDialog, !app::settings::get<bool>("open_save", "native_dialog"));
 #endif
 
         if ( dialog.exec() == QDialog::Rejected )